requirements.txt
coinbase-advanced-py
discord.py
